create separate indexedDBService (as raw api) and move app logic to storageService level
