+++ src/services/indexedDBService.ts
@@ -1,0 +1,38 @@
+import type { DailyRecord } from '../types';
+
+const DB_NAME = 'lunaflow';
+const DB_VERSION = 2;
+const STORE_NAME = 'appData';
+const STORE_KEY = 'events';
+
+const openDB = (): Promise<IDBDatabase> =>
+  new Promise((resolve, reject) => {
+    const request = indexedDB.open(DB_NAME, DB_VERSION);
+    request.onupgradeneeded = () => {
+      if (!request.result.objectStoreNames.contains(STORE_NAME)) {
+        request.result.createObjectStore(STORE_NAME);
+      }
+    };
+    request.onsuccess = () => resolve(request.result);
+    request.onerror = () => reject(request.error);
+  });
+
+export const readDailyRecords = async (): Promise<DailyRecord[] | null> => {
+  const db = await openDB();
+  return new Promise((resolve, reject) => {
+    const tx = db.transaction(STORE_NAME, 'readonly');
+    const request = tx.objectStore(STORE_NAME).get(STORE_KEY);
+    request.onsuccess = () => { db.close(); resolve(request.result ?? null); };
+    request.onerror = () => { db.close(); reject(request.error); };
+  });
+};
+
+export const writeDailyRecords = async (records: DailyRecord[]): Promise<void> => {
+  const db = await openDB();
+  return new Promise((resolve, reject) => {
+    const tx = db.transaction(STORE_NAME, 'readwrite');
+    const request = tx.objectStore(STORE_NAME).put(records, STORE_KEY);
+    request.onsuccess = () => { db.close(); resolve(); };
+    request.onerror = () => { db.close(); reject(request.error); };
+  });
+};
